requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
black==24.8.0
click==8.1.8
django==4.2.23
django-allauth==65.9.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mypy-extensions==1.1.0
mysqlclient==2.2.7
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.6
PyJWT==2.9.0
sqlparse==0.5.3
tomli==2.2.1
typing-extensions==4.13.2